+++ frontend/src/components/dialog/export/ExportWord.tsx
@@ -1,6 +1,6 @@
 import React, {useContext} from 'react';
 import { Button } from "antd";
-import {MyIcon} from "@/components/Menu";
+import { FileWordOutlined } from '@ant-design/icons';
 import useProjectStore from "@/store/project/useProjectStore";
 import shallow from "zustand/shallow";
 import { ProjectMenuCloseContext } from "@/components/Menu/projectMenuClose";
@@ -15,7 +15,7 @@ const ExportWord: React.FC = () => {
       type="text"
       size="small"
       block
-      icon={<MyIcon type="icon-file-word"/>}
+      icon={<FileWordOutlined />}
       onClick={() => {
         closeProjectMenu();
         projectDispatch.exportFile('Word');